pyproject.toml
[project]
name = "chatty"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "absl-py>=2.3.1",
    "asyncio>=4.0.0",
    "datasets>=4.1.1",
    "openai>=2.0.1",
    "pandas>=2.3.3",
    "playwright>=1.55.0",
    "pyaml>=25.7.0",
    "pymupdf>=1.26.4",
    "together>=1.5.26",
    "tqdm>=4.67.1",
]
